feat: remove show snackbar function
This commit removes the `showSnackbar` function from the `AuthCubit` and moves it to the `CoreUtils`.

The `showSnackbar` function is a utility function that displays a snackbar on the screen. It is used in the `AuthCubit` to display errors and messages to the user. However, the `showSnackbar` function is not specific to the `AuthCubit` and can be used in any part of the application.

By moving the `showSnackbar` function to the `CoreUtils`, it can be reused in any part of the application without having to duplicate the code. This makes the code more DRY and easier to maintain.